requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
decorator==5.1.1
Django==4.1
django-summernote==0.8.20.0
django-taggit==3.0.0
executing==0.8.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.7.1
flake8==4.0.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
jieba3k==0.35.1
joblib==1.1.0
lxml==4.9.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mysql-connector-python==8.0.30
newspaper3k==0.2.8
nltk==3.7
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.7.9
requests==2.28.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
tinysegmenter==0.3
tldextract==3.3.1
tqdm==4.64.0
traitlets==5.2.2.post1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
